fix: gitlab is missing the board entity
